Test case for all permitted hours worked and implementation
Added comments to clarify the branching logic at each step as the 24 hour clock can make it confusing. Also removed magic number of 17 into a const called FivePM. Added an if check at beginning to verify start and endTime are not the same value, may need to adjust this later. All tests green.